pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "qPrimer"
version = "1.0.3"
authors = [{ name = "Xiaodong Li", email = "lxd1997xy@163.com" }]
description = "A library for qPCR primers design, check, annotate and visualize."
readme = "README.md"
keywords = ["qPCR", "design primers", "check primers", "annotate primers", "visualize primers"]
requires-python = ">=3.9"
license = { file = "LICENSE" }
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.9",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    'setuptools>=65.6.3',
    'biopython>=1.81',
    'numpy>=1.24.1',
    'pandas>=1.3.3',
    'primer3-py>=2.0.3',
    'jinja2>=3.0.2'
]
dynamic = ['entry-points']

[project.urls]
Homepage = "https://github.com/swu1019lab/qPrimer"
issues = "https://github.com/swu1019lab/qPrimer/issues"

[project.scripts]
qPrimer = "qPrimer.qPrimer:main"

